+++ ptahlmud/backtesting/portfolio.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass
 from datetime import datetime
 from decimal import Decimal
 
-from ptahlmud.backtesting.exposition import Trade
+from ptahlmud.backtesting.trading import Trade
 
 
 @dataclass(slots=True)
