+++ tests/backtesting/test_trading.py
@@ -7,7 +7,7 @@ from hypothesis import given
 from hypothesis import strategies as some
 from hypothesis.strategies import composite
 
-from ptahlmud.backtesting.exposition import Position
+from ptahlmud.backtesting.trading import Position
 from ptahlmud.types.signal import Side
